Fix implementation of ReqUserSearchSet
