+++ MHTriServer/Server/Packets/ReqUserSearchSet.cs
@@ -1,43 +1,43 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
 using System.Diagnostics;
 
 namespace MHTriServer.Server.Packets
 {
-    public class ReqUserStatusSet : Packet
+    public class ReqUserSearchSet : Packet
     {
-        public const uint PACKET_ID = 0x66400100;
+        public const uint PACKET_ID = 0x66300100;
 
-        public List<UnkByteIntStruct> UserStatus { get; private set; }
+        public List<UnkByteIntStruct> UnknownField { get; private set; }
 
-        public ReqUserStatusSet(List<UnkByteIntStruct> userStatus) : base(PACKET_ID) => (UserStatus) = (userStatus);
+        public ReqUserSearchSet(List<UnkByteIntStruct> unknownField) : base(PACKET_ID) => (UnknownField) = (unknownField);
 
-        public ReqUserStatusSet(uint id, ushort size, ushort counter) : base(id, size, counter) { }
+        public ReqUserSearchSet(uint id, ushort size, ushort counter) : base(id, size, counter) { }
 
         public override void Serialize(ExtendedBinaryWriter writer)
         {
             base.Serialize(writer);
-            UnkByteIntStruct.SerializeArray(UserStatus, writer);
+            UnkByteIntStruct.SerializeArray(UnknownField, writer);
         }
 
         public override void Deserialize(ExtendedBinaryReader reader)
         {
             Debug.Assert(ID == PACKET_ID);
-            UserStatus = UnkByteIntStruct.DeserializeArray(reader);
+            UnknownField = UnkByteIntStruct.DeserializeArray(reader);
         }
 
         public override string ToString()
         {
-            var str = $":\n\tUserStatus({UserStatus.Count})";
-            for (var i = 0; i < UserStatus.Count; ++i)
+            var str = $":\n\tUnknownField({UnknownField.Count})";
+            for(var i = 0; i < UnknownField.Count; ++i)
             {
-                var data = UserStatus[i];
+                var data = UnknownField[i];
                 str += $"\n\t  [{i}] =>\n\t    UnknownField {data.UnknownField}";
                 if (data.ContainUnknownField3)
                 {
-                    str += $"\n\t    UnknownField3 {data.UnknownField3}";
+                    str += $"\n\t    UnknownField2 {data.UnknownField3}";
                 }
             }
-
             return base.ToString() + str;
         }
     }
